List plpythonu requirement first, so get pg_catalog scanned before public
Since "plpythonu" is installed in "pg_catalog" schema, requiring it
from cartodb.control adds the "pg_catalog" to the search_path.
Schemas are added to search_path in order of appearence in the "requires"
directive of the cartodb.contorl, right after the name of schema
written in the "schema" directive.

So this commit changes the resulting search_path from:

 cartodb,schema_triggers,public,pg_catalog

To:

 cartodb,pg_catalog,schema_triggers,public

Preventing presence of objects in the "public" schema from changing
interpretation of function signatures and body from this extension.

Spotted in the real world in presence of a "date" table changing
intepretation of CDB_StringToDate function.