Use stddev_pop for stats
stddev (stddev_samp) has a correction factor for only sampling a
portion of the population. Because we are sampling the entire array
stddev_pop is more appropriate. This allows a decrease in the
allowable error in the tests.

The accumlated error for kurtosis is approximately 2e-8 and for
skewness no error is observed for the test data.