COALESCE source columns
What it does: it changes the generated UPDATE SET caluse so that
instead of

  the_geom = changed.the_geom

it will write

  the_geom = COALESCE(changed.the_geom, the_geom)

So, if the source column has a NULL value, it will keep whatever value
the target column has for that row.

This may be a general mechanism to work around syncing in a general
case. We may decide to apply it to specific columns instead but that
complicates the implementation (instead of skip_columns it may have
a harder-to-understand param "coalesce_columns", *if* they exist).

I'm creating this to discuss it. Note it greatly affects the simple
API.