Change type of server_alias from text to name
On the one hand this is done for consistency with the rest of
parameters.

On the other hand this is the "correct" type, as it is more
restrictive than `text` type (`name` is limited to 64 bytes and the
generated object names will use that type and suffer the same length
limitation) and it is more suitable to build db object names.