Refactor: remove overloaded CDB_SetUp_PG_Federated_Table methods
In order to do so, the main method has to have a couple of default
values for input params, and perform a number of null checks.

The advantage is that it reduces the code by 1/2 and that fixes are to
be performed in just one place, insteda of spreading it 2 functions.